nothing

+++ InputFile.h
@@ -1,3 +1,4 @@
+
 #pragma once
 #include "CGraph.h"
 #include "Utils.h"
@@ -8,7 +9,7 @@ using namespace DS;
 
 class InputFile{
     public:
-    void readInput(string path,CGraph<Place,double> graph){
+    void readInput(string path,CGraph<Place,float> graph){
         ifstream in;
         in.open(path);
         string s,line;
